+++ source/kdtree_constants.h
@@ -34,7 +34,7 @@ struct Constants {
         // calculate distance between points of different
         // cardinality
 
-    static const double KDTREE_ERROR_INDEX;
+    static const size_t KDTREE_ERROR_INDEX;
         // Denotes an error output resulting from an attempt to
         // search for a point either on an empty tree on in case of
         // cardinality mismatch
